Consolidated the headers used for the common/ code to call back to each architecture-specific installer. These are moved together under `<local/*>` per architecture.
svn path=/trunk/; revision=3171

+++ trans/src/hppa/local/localtypes.h
@@ -11,18 +11,20 @@
 #define LOCALTYPES_INCLUDED
 
 #include <construct/installtypes.h>
-#include "exptypes.h"
+#include <local/exptypes.h>
 
 typedef struct {
-    int ashsize;
-    int ashalign;
+    long ashsize;
+    long ashalign;
 } ash;
 
 extern ash ashof(shape);
+
 
 typedef struct {
     bool is_special;
     exp special_exp;
 } speci;
+
 
 #endif /* LOCALTYPES_INCLUDED */
