+++ trans/src/sparc/local/localtypes.h
@@ -11,20 +11,18 @@
 #define LOCALTYPES_INCLUDED
 
 #include <construct/installtypes.h>
-#include "exptypes.h"
+#include <local/exptypes.h>
 
 typedef struct {
-    long ashsize;
-    long ashalign;
+    int ashsize;
+    int ashalign;
 } ash;
 
 extern ash ashof(shape);
-
 
 typedef struct {
     bool is_special;
     exp special_exp;
 } speci;
-
 
 #endif /* LOCALTYPES_INCLUDED */
